requirements.txt
paho-mqtt
pyJWT
google-cloud-vision
lxml
